Add Hibernate ORM to project JPA/Criteria API

+++ computer-database/src/com/excilys/cdb/controller/cli/ComputerController.java
@@ -17,7 +17,7 @@ public class ComputerController {
 		this.computerService = computerService;
 	}
 	
-	public int getTotal() {
+	public Long getTotal() {
 		return this.computerService.getTotal();
 	}
